Migrate Module to separate file
